Configuration

+++ Modules/Stream/database/migrations/2025_04_22_181909_create_streams_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('language_level_id')->constrained()->onDelete('cascade'); // (A1–C2)
             $table->foreignId('teacher_id')->constrained('users')->onDelete('cascade');
-            $table->unsignedInteger('current_subject_number')->default(1);
+            $table->unsignedTinyInteger('current_subject_number')->default(1);
             $table->enum('status', ['planned', 'started', 'paused', 'finished'])->default('planned'); // status
             $table->date('start_date');
             $table->date('end_date')->nullable();
